(re) Moves extend helper to a separate module
